Move wgts/common/cnv into cnv_purple as legacy tools/constants
