+++ src/lib/djerba/plugins/wgts/cnv_purple/legacy_constants.py
